moving optimization tests to dedicated folder
with the exception of the z_* tests, because they need to run after the
setup test
