+++ tests/optimizations/cube_stars_optimization_test.py
