+++ tests/optimizations/encounters_optimization_test.py
