+++ tests/optimizations/shock_inputs.csv
